@@ -1,9 +1,11 @@
-package com.maaz.interiar;
+package com.maaz.interiar.ui.activities;
 
 import android.content.Intent;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.maaz.interiar.R;
 
 public class MainActivity extends AppCompatActivity {
 
