@@ -1,14 +1,14 @@
-package com.maaz.interiar;
+package com.maaz.interiar.ui.fragments;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class cartfragment extends Fragment {
+import com.maaz.interiar.R;
+
+public class CartFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
