@@ -1,12 +1,14 @@
-package com.maaz.interiar;
+package com.maaz.interiar.ui.fragments;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.maaz.interiar.R;
 
 public class Profilefragment extends Fragment {
     @Nullable
